Updated index.ts imports for forms
