+++ src/Forms/Form/Col.tsx
@@ -1,7 +1,7 @@
 import * as React from "react";
 import { memo, useMemo } from "react";
 import styled from "styled-components";
-import { actions } from "./formUtils";
+import { formActions } from "./formActions";
 
 import FormFieldText from "../FormFieldText";
 import FormFieldTextArea from "../FormFieldTextArea";
@@ -70,7 +70,7 @@ const Col = (props: ColPropsTypes) => {
 		return fieldsDef.reduce((prev, curr) => {
 			prev[curr.name] = async function (value) {
 				await dispatch(
-					actions.setFieldValue({
+					formActions.setFieldValue({
 						name: curr.name,
 						value,
 					})
@@ -85,13 +85,13 @@ const Col = (props: ColPropsTypes) => {
 		return fieldsDef.reduce((prev, curr) => {
 			prev[curr.name] = async function () {
 				await dispatch(
-					actions.validateField({ name: curr.name })
+					formActions.validateField({ name: curr.name })
 				);
 
 				if (curr.pairedFields)
 					curr.pairedFields.forEach(async pairedField => {
 						await dispatch(
-							actions.validateField({ name: pairedField })
+							formActions.validateField({ name: pairedField })
 						);
 					});
 			};
