+++ src/Forms/FormFieldImageVideoLinkDocumentBrowsing/ImageVideoLinkDocumentBrowsing.test.tsx
@@ -8,7 +8,7 @@ import Form from "../Form/Form";
 
 // Utils
 import { menuOptions } from "../MenuFormFieldCard/MenuFormFieldUtils";
-import { actions, useForm } from "../Form/formUtils";
+import { formActions, useForm } from "@root/forms/Form";
 import {
 	documentExample,
 	imageAssetExample,
@@ -36,7 +36,7 @@ const ImageVideoLinkDocumentBrowsingExample = () => {
 	const setImage = async () => {
 		setImageCallback();
 		await dispatch(
-			actions.setFieldValue({
+			formActions.setFieldValue({
 				name: "imageVideoLinkDocumentBrowsing",
 				value: imageAssetExample,
 			})
@@ -46,7 +46,7 @@ const ImageVideoLinkDocumentBrowsingExample = () => {
 	const setVideo = async () => {
 		setVideoCallback();
 		await dispatch(
-			actions.setFieldValue({
+			formActions.setFieldValue({
 				name: "imageVideoLinkDocumentBrowsing",
 				value: videoAssetExample,
 			})
@@ -55,7 +55,7 @@ const ImageVideoLinkDocumentBrowsingExample = () => {
 
 	const setDocument = async () => {
 		await dispatch(
-			actions.setFieldValue({
+			formActions.setFieldValue({
 				name: "imageVideoLinkDocumentBrowsing",
 				value: documentExample,
 			})
@@ -65,7 +65,7 @@ const ImageVideoLinkDocumentBrowsingExample = () => {
 
 	const setLink = async () => {
 		await dispatch(
-			actions.setFieldValue({
+			formActions.setFieldValue({
 				name: "imageVideoLinkDocumentBrowsing",
 				value: linkExample
 			})
@@ -75,7 +75,7 @@ const ImageVideoLinkDocumentBrowsingExample = () => {
 
 	const handleRemove = () => {
 		dispatch(
-			actions.setFieldValue({
+			formActions.setFieldValue({
 				name: "imageVideoLinkDocumentBrowsing",
 				value: [],
 			})
