+++ src/index.ts
@@ -14,11 +14,8 @@ import Spinner from "./components/Spinner.jsx";
 export * from "./transforms";
 export * from "./types";
 
-export { default as FormFieldAddress } from "./forms/FormFieldAddress";
-export * from "./forms/FormFieldAddress";
-
-export { default as FormFieldAdvancedSelection } from "./forms/FormFieldAdvancedSelection";
-export * from "./forms/FormFieldAdvancedSelection";
+export { default as Form } from "./forms/Form";
+export * from "./forms/Form";
 
 export { default as FilterDate } from "./components/FilterDate";
 export * from "./components/FilterDate";
@@ -41,38 +38,11 @@ export * from "./components/Chip";
 export { default as CheckboxList } from "./components/CheckboxList";
 export * from "./components/CheckboxList";
 
-export { default as FormFieldColorPicker } from "./forms/FormFieldColorPicker";
-export * from "./forms/FormFieldColorPicker";
-
-export { default as DatePicker } from "./forms/FormFieldDate/DatePicker";
-export * from "./forms/FormFieldDate/DatePicker";
-
 export { default as IconButton } from "./components/IconButton";
 export * from "./components/IconButton";
 
 export { default as FormDialog } from "./components/Dialog";
 export * from "./components/Dialog";
-
-export { default as FormFieldCheckbox } from "./forms/FormFieldCheckbox";
-export * from "./forms/FormFieldCheckbox";
-
-export { default as FormFieldImageUpload } from "./forms/FormFieldImageUpload";
-export * from "./forms/FormFieldImageUpload";
-
-export { default as FormFieldPhoneSelectionDropdown } from "./forms/FormFieldPhoneSelectionDropdown";
-export * from "./forms/FormFieldPhoneSelectionDropdown";
-
-export { default as FormFieldRadio } from "./forms/FormFieldRadio";
-export * from "./forms/FormFieldRadio";
-
-export { default as FormFieldChipSingleSelect } from "./forms/FormFieldChipSingleSelect";
-export * from "./forms/FormFieldChipSingleSelect";
-
-export { default as FormNav } from "./forms/FormNav";
-export * from "./forms/FormNav";
-
-export { default as FormFieldImageVideoLinkDocumentBrowsing } from "./forms/FormFieldImageVideoLinkDocumentBrowsing";
-export * from "./forms/FormFieldImageVideoLinkDocumentBrowsing";
 
 export { default as LeftNav } from "./components/LeftNav";
 export * from "./components/LeftNav";
@@ -89,29 +59,11 @@ export * from "./components/MenuSelect";
 export { default as Modal } from "./components/Modal";
 export * from "./components/Modal";
 
-export { default as FormFieldText } from "./forms/FormFieldText";
-export * from "./forms/FormFieldText";
-
-export { default as FormFieldTable } from "./forms/FormFieldTable";
-export * from "./forms/FormFieldTable";
-
-export { default as FormFieldTextArea } from "./forms/FormFieldTextArea";
-export * from "./forms/FormFieldTextArea";
-
-export { default as TimePicker } from "./forms/FormFieldDate/TimePicker";
-export * from "./forms/FormFieldDate/TimePicker";
-
 export { default as ToggleSwitch } from "./components/ToggleSwitch";
 export * from "./components/ToggleSwitch";
 
-export { default as TopComponent } from "./forms/TopComponent";
-export * from "./forms/TopComponent";
-
 export { default as RadioButton } from "./components/RadioButton";
 export * from "./components/RadioButton";
-
-export { default as FormFieldDate } from "./forms/FormFieldDate/DateField";
-export * from "./forms/FormFieldDate/DateField";
 
 export { default as Snackbar } from "./components/Snackbar";
 export * from "./components/Snackbar";
@@ -147,6 +99,3 @@ export {
 	Menu,
 	Spinner
 }
-
-export { default as DropdownSingleSelection } from "./forms/FormFieldDropdownSingleSelection";
-export * from "./forms/FormFieldDropdownSingleSelection";
